requirements.txt
web3
requests
numpy
python-binance
pandas
scikit-learn
alpaca-trade-api
mkdocs
mkdocs-material
python-dotenv
